Fix issue preventing spectacle from opening
Summary:
Commit dbb1600ce2e09f8d9191b7923a667bf3ea55fdcd attempted to unify the way how
copy and save are handled. However there was one oversight. On image copy the
check whether to quit doesn't happen on button press but after a image is copied.
This includes when the copy is triggered automatically. If the user has automatic
copying of screenshots enabled, spectacle will not open again because upon start
the automatically taken screenshot is copied and spectacle terminates itself.
This introduces a `copy` method similiar to `save()` and `saveAs()` which does
the check and triggers copying.

Test Plan:
Enable automatic copying of screenshots and check quit after save or copy.
Close Spectacle.
Open Spectacle.

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D25765