Find qdbus command correctly
Summary:
Spectacle's Desktop file assumes that `qdbus` is in $PATH. However this is not
guaranteed; the command lives in the Qt binaries dir which is not typically in $PATH,
and distros typically create symlinks with different names in `/usr/bin`, but not all
create a second compatibility symlink at `/usr/bin/qdbus` that points to the version
for the current Qt version.

Therefore, we should find the command in the Qt binaries dir itself rather than relying
on these compatibility symlinks.

BUG: 413007
FIXED-IN: 20.04.0

Test Plan:
- Delete the `/usr/bin/qdbus` symlink, if you have one
- Compile and install Spectacle
- Hit PrintScreen
- See that Spectacle still opens

Reviewers: davidedmundson, #spectacle, sitter

Reviewed By: sitter

Subscribers: sitter, #plasma, rdieter, kossebau, fvogt, davidre

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D25883