Fix capturing from depth 30 screens (10bpc)
The color order was incorrect, or rather, not matching the screen
created by the amdgpu DDX driver.

There's probably a better and more portable way to create a compatible
QImage for an XCB image, but I don't know what it is. At least now the
color order for 30bpp matches 24bpp, so on a given platform it should be
either always correct or always wrong.

BUG: 422285