fix autoincrement of %d for templates with subdirs
in case the filename to save a screenshot to contains subdirectories,
the auto-increment of any %d placeholders was broken.

For a template of '%Y/%M/screen-%d' and '%Y/%M/screen-1.png' already existing
this now results in '%Y/%M/screen-2.png' being created instead of '%Y/%M/screen-1-1.png'.