We do not need progress info when checking if a file exists
I debugged the time loss of
BUG:430173
to be inbetween when org.kde.JobViewServer.requestView is called inside
KJobWidgets and the DBus adaptor on plasmashell is notified of that.
Unfortunately our methods are implemented in a blocking way as is the call in
KUiServerJobTracker::registerJob. So a simple fix is to disable job tracking.
Furthermore if the job would take long enough to display progress, it would
display an error to the user if the file doesn't exist, which is clearly not
what we want.
FIXED-IN:20.12.1