Don't translate property name
kcfg_property is a magic property used by KConfigDialog to control what property is manipulated.
In this case we want to manipulate the currentText property of the QComboBox.

Strings in UI files are getting translated by default. This results in kcfg_property containing a nonexistant property for non-english locale which breaks setting the right value of the combobox.

BUG: 431557