feat: use radio buttons for after screenshot clipboard actions
Move UI elements, and remove unnecessary Save location
radiobutton code

fix: disable 'copy to clipboard' checkbox when 'copy file location' is also selected

spectacle previously allowed copying both the image as well as file
location simultaneously to the clipboard. This MR also attempts to fix a corner case which
requires 'Autosave image' to be enabled to 'copy file location to clipboard'.
Previously, if 'Autosave Image' checkbox was not enabled, it was still possible for the user
to copy the file location to clipboard. However, this had indirectly triggered an autosave first.
This is confusing, as spectacle saves the file to its default save location without it
being asked to. So it would be ideal to disable 'Copy file location' if 'Auto Save' option was
not enabled in the first place

fix: typo in label name