feat: add support for copying saved screenshot file path to clipboard
this feature aims to enable 'Copy File Location' to clipboard,
this allows users to get a direct local file path to the image if
the user has enabled 'Auto Save' option. If 'Auto Save' is disabled, a
temporary screenshot file will be created, and that file path would
be copied to clipboard.

This feature is particularly useful for users, who would directly like to
upload their screenshot by file path using a FileChooser, eg: Browser,
or to upload the screenshot to non-KServices, like a private cloud instance, etc.