Simplify branching for the notify text and fix messages containing dots
Simplify the logic by removing the bool isBackgroundMode. Fix double notifications
where one has dots instead of the filename and save path.

Add missing notification text if mCopyLocationToClipboard is set. Uses 2in1
principle: Considering that when mCopyLocationToClipboard is true it means that
the image also gets saved as a file, so this makes the notification text show
both the path where it's saved & the name of it _and_ the information that the
file path has been added to the clipboard.

BUG: 447517
FIXED-IN: 22.04