Do not apply smoothing to the magnifier's image
Problem: During the selection of the area to screenshot with
"Rectangular Region", we get an image that was smoothed with bilinear
interpolation, but it makes it difficult to create pixel-perfect
selections as you can't properly see what's being selected.

Solution: Do not setting antialiasing or smoothing to the magnifier's
image. Note that I also remove antialiasing because the crosshair drawn
in the magnifier would appear to be slightly misaligned with the image's
pixels, and this is undesirable.