Display shortcuts in capture mode combobox
Shows the global shortcuts in the capture mode ComboBox. These are
the shortcuts that would perform a screenshot in that capture mode
directly even if Spectacle wasn't currently open.

The purpose of this commit is to teach users these shortcuts,
similar to how a QMenu does it, to boost users' speed and
productivity long term.

Part of the credit for this MR goes to Leon De Andrade who uploaded
a Phabricator revision to the same effect three years ago:
https://phabricator.kde.org/D23446

BUG: 389778