Hide "On click" checkbox instead of disabling it
When this feature isn't supported, currently its checkbox gets disabled.
Because there is no way to make it enabled without rebooting into an
X11 session, its permanently disabled presence is frustrating to the
user and consumes visual space.

To fix this, hide the checkbox instead of disabling it when its feature
is not available.