ImageCaptureOverlay: refine toolbar positioning, allow toolbar dragging
Center without selection, put along the bottom of the inner area of the
selection when the selection is large enough that the toolbar has to fit
inside.

Allow dragging main toolbar and annotation tools toolbar around when
there's no selection and remember positions unless the selection
changes. I think this is needed because it's impossible to know which
part of the screen the user will want to annotate when there's no
selection. Not allowing dragging with a selection because the toolbars
for 2 screens have to act like they're one toolbar when a selection
spans 2 screens. That's hard to deal with when allowing them to be dragged.