Drop plasmashell wayland protocol usage
In the past QuickEditor used one giant window across all screens
so it was needed to be able to position it and go above everything.
Now we use one window for each screen so xdg_toplevel set_fullscreen
should work just as well.
This allows us to drop KWayland dependency.