Make CaptureWindows with unique_ptr, move list of windows to SpectacleWindow
unique_ptr will help prevent leaks when SpectacleCore is destroyed.

Having the list of windows in SpectacleWindow makes it possible to add
an remove windows from the constructor and destructor rather than
manually syncronizing them with m_captureWindows and m_viewerWindow in
SpectacleCore.

Also move connections and event filter installation for SelectionEditor
and CaptureWindow to CaptureWindow's constructor and destructor. We can
get rid of captureWindowAdded and captureWindowRemoved and remove the 2
connections for handling those 2 signals.