Only UI CaptureModes, no ExportManager CaptureMode, better CLI handling
This is all one commit since the changes require each other to keep
Spectacle functioning as well as it did before the patch.

ExportManager only stored the capture mode for determining whether to
format window titles in file names. This change
makes it so we don't have to syncronize as many states with
ExportManager.

PlatformXcb now clears ExportManager's window title when a non-window
grabbing GrabMode is used via the windowTitleChanged signal.

CaptureModeModel now only has capture modes represented in the UI. For
getting the TransientWithParent grab mode, we are supposed to use the
transientOnly setting. Using that is awkward, but that's how it was
supposed to be and would take a lot more effort to change in a way that
has a high chance of being genuinely better.

Command line options are now more organized in the code and more
consistently handled in actual use. When running spectacle via CLI with
an existing instance and without specifying --new-instance, any command
line options you specify will now be respected in the existing instance
instead of just being ignored or only partially used.