VideoPlatform: Make the value of recordedTime update more reliably
The value didn't always update as consistently and frequently as it
should have. It's still not completely perfect since sometimes the
duration of the finished video is slightly different from the recorded
time indicated while recording.

Switched from using QDateTime to QElapsedTimer because using QDateTime
to get the time since something happened can be problematic if the user
changes their computer's time.

Used a QBasicTimer instead of a QTimer because it's slightly more
efficient and not really any more work in this case.