Minimize Spectacle while recording if intersecting target
If Spectacle's window geometry reported by QWindow::frameGeometry()
intersects the area we are recording, we minimize Spectacle. The
exception is when recording Spectacle itself.

I couldn't find a foolproof way to deal with window geometry on wayland,
so hopefully this is good enough. The geometry reported by
QWindow::frameGeometry() is not necessarily the same as what is reported
by queryWindowInfo, which comes from KWin's own Window::frameGeometry().