Don't close all windows if activated by DBus with GUI and no CLI options
No options and having a GUI is a heuristic for activating Spectacle via
the PrintScreen key. Maybe we can find a better way to determine this
later, but I'm trying to keep up with the release schedule for now.

BUG: 467118


(cherry picked from commit 068c6f548688b3425a94786cacadbf5302a6e123)