Use export action flags instead of several functions and signals
This allows us to specify export actions and figure out which kinds of
actions have been done simultaneously without having separate functions
and signals for all the various combinations of save, save as, copy
image and copy path.

SpectacleCore: Instead of storing which actions to use in boolean member
variables, use a function to calculate which automatic export options
can be done. This makes it so we don't need to update the variables
every time we want to use them.

CaptureWindow, SelectionEditor, SpectacleCore: Don't export directly
from CaptureWindows. Instead, pass export actions to the
SpectacleCore::grabDone() signal via SelectionEditor::acceptSelection().
Necessary for preventing double saves when clicking the Save button in
rectangle capture windows.

In order to cherry-pick this to 23.04, we'll need to cherry-pick some of
the previous code clean-up and bug fix commits on master:

- 35f3eefc2125aadd302519bee629fa380e0d1cce
- 0bf9cd3433e418ce15337490cf119e0e30f87ee6
- 6711242f8397417ae05eaf76dc5f81622b691473
- 3a9f5351deb887eeac5d751d44fa55e8c31e63d4
- a95993c25059ab8de85f7128ec87b32f6c7df9b4
- ed981b997cfc0d155db29af4bf7f04659903092e
- 28f63cee4fb71de2804994b582b084db77b2cfcc
- 211127024a500da6f9ccf27c1131682a5558eec0
- 553eb2cf16ad3ccb8f9177cfd0ec533de5f22fdc
- 0fda35cb61b8db8cc50e05e28ee33e7516ca7cd2
- 001877540c54c19ba0a08b34182744d285dac3bc
- abf044557f36115c8e1ace0ba86a5c020422288f

CCBUG: 422874 (needs new translatable strings to fully fix)