Clean up edit existing option code and allow overwriting the image
The edit existing image options' behavior was written in an
unnecessarily complicated way for the current design of the app. Now,
most of it is handled in SpectacleCore::activate().

It is now possible to edit an image and have it overwrite the original
one when you press Save in the GUI.

There is also logic for clearing the edit existing option when taking
a new screenshot via SpectacleCore::takeNewScreenshot() to prevent
taking new screenshots from overwriting the edited existing image.