Fix quitting Spectacle with Escape affecting windows below it
Previously, if you pressed Escape to close Spectacle when a fullscreen
video in Firefox was underneath, the release of Escape would also cancel
the fullscreen state of the video. Basically, Spectacle was no longer
there to accept the Escape release event, so it went to Firefox.

I think this is fundamentally an upstream issue in Qt, KWin or X11
(intentional or not), but we're fixing it here because we can. What and
where the correct upstream behavior should be is also a complex and
highly debatable topic.

BUG: 428478


(cherry picked from commit 2af57799757a8f472a3b7615adeb5336d3bed292)