Remove support for KWin Wayland 5.20 and lower
It's annoying to maintain 2 platform abstractions for KWin Wayland. The
previous LTS KWin version was 5.24 and the current LTS version is 5.27.
I don't think we need to support even older versions of KWin anymore and
we're already working on moving to version 6.

I haved renamed PlatformKWinWayland2 to PlatformKWinWayland since we no
longer have 2 versions.

This change is intended for version 23.08, not 23.04.