Add hover outline for selectable annotation objects
There is now a text color/background color dashed outline around the
visualGeometry of hovered visible EditActions. I implemented it using a
QML item instead of in the paint event because it repaints faster at
high zoom levels that way. It might be a bit troublesome to change if we
ever want to switch to more tightly fitting outlines instead of simple
rectangle outlines.