Hide redundant header on Wayland
On Wayland where screen recording is supported, the right column has a
tab bar on top with the text "Screenshot" and "Recording" for the tabs.

On each of the tabs, right below that tab bar is a header that says a
wordier version of the same thing, which is rather redundant.

Hide the labels on Wayland when the tab bar is visible.

BUG: 468818
FIXED-IN: 23.04.1


(cherry picked from commit 9ef2dcdf477020fd2ba8ecb4f2678e097c75b365)