PlatformKWin: use logical screen geometry on X11
On X11 QScreen::geometry has a scaled and rounded size, but a raw
position. We divide screen position and image size by the DPR on X11 to
get the logical (wayland) geometry.

For some reason, the "scale" property from the DBus screenshot metadata
is 1 when it shouldn't be. That will need to be fixed, but for older
versions of KWin, we can at least use the screen DPR. The scale property
and screen DPR should be the same.

We also assume that all screens on X11 have the same DPR. Plasma System
Settings only allows global scaling on X11 and this makes the math
easier.