Fix rectangle screenshots with less than 100% UI scale
Fixes it by just doing a normal all screens grab instead of trying to
get an image for each screen and then trying to position and scale them
all correctly for the given platform. This should also be a bit more
efficient.

Also lays some of the ground work for rectangle video recording by
making SelectionEditor send a rectangle and export actions instead of an
image and export actions.

BUG: 462860
BUG: 473888
FIXED-IN: 23.12.0