Save image as temp file, then copy data when copying image
Now you can see a thumbnail of the image you copied in a notification
when only copying an image. The image is in whatever format you just
saved (if saving and copying image) or the default image format.

There is a known minor flaw where if you click the image thumbnail to open it in an image viewer, the image will be deleted before it can be viewed. This is because the image is in the temp dir, so it gets deleted when spectacle is closed.

BUG: 422874
BUG: 465781