Change filename placeholder format
Use `<placeholder>` instead of `%placeholder`. This makes it easier to
use
more descriptive or complex placeholders.

Also use more descriptive placeholders since we can and migrate old
placeholders.

With this new format we could add things like `<ISODate>` or
`<ShortDate>` as shorthands for common date/time formats.

I use `<key>` instead of `%key%` because sequences with obvious opening
and closing characters are easier to parse for humans and machines. I
chose '<' and '>' instead of other pairs of characters because I don't
need to escape '<' or '>' in regex. I do need to escape them in Rich
Text/HTML/XML, so it's sometimes less convenient than '{' and '}', but
not too bad.

BUG: 476023