Show magnefier with Shift and use magnifier locations
You can now get the magnifier while using arrow keys to modify the
selection or hovering by holding Shift. The magnifier tries to keep its
location close to where the mouse was, the last handle you dragged with
the mouse or the corner closest to the last part of the selection that
was modified.

The magnifier is only shown now while Shift is held and the "Show
magnifier" setting is enabled, but I'm not actually that happy with this
behavior even if it seems more accurate to what the UI says. I think I
want to make the setting toggle between always showing while modifying
the selection and only showing while shift is held.

BUG: 477399