Simplify arrow key magnifier location
I think the previous behavior was unnecessarily complex and slightly
annoying to predict. The original idea was that corners were better for
showing magnification on since where they are relative to the boundaries
of the selection is more obvious. I think if I want to be able to put
the magnifier in a corner, I should find a way to detect when multiple
arrow keys are held.