Change "Show less" to "Show fewer" since the thing in question is countable
"Less" is used for nouns of indefinite quantity. For example,
"There is less rice." In this case "rice" is a noun of indefinite
quantity; you don't say "there are three rices."

For quantities that are definite or countable, you use "fewer". For
example, "There are fewer *grains* of rice" since "grains" is countable.

In this case, the things in the list are defined and countable, so we
should say "Show fewer".