Fix recording
Make sure we don't enter the recording state until all the properties
are set.

When the code was originally implemented, we didn't have the blocking
call to set up the temp path, this makes it possible for the
ScreencastingStream::created signal to emit while we are at
::setupOutput.

To address that, guard the activation to make sure both values are on
point.