Initial implementation of scanning captured screenshot for qrcodes.
The scanner automatically scans the captured image.
If a qr code is found, it can let you open the link / copy it's text to clipboard via. the notification.
If multiple QR Codes are found, it shows the first QR Code. This is a design decision to make things less confusing/cluttered. The user can always capture a specific area of the screen to fetch the intended QR Code.