Traits: Remove type() from Fill and Text and make Text itself a variant
type() wasn't really doing anything to ensure type safety since we're
not using class enums. We're not using class enums so we can use enums
with std::get<index> to avoid saying the exact type. The exact type
doesn't necessarily tell you the meaning of what you're getting.

Making Text itself a variant instead of having a value member var that
is a variant because its slightly more convenient and true to the
meaning of Text.