AnnotationViewport: Don't use TextureCanUseAtlas
It was the wrong thing to use with these textures and was likely causing the GPU to run out of memory. That caused the textures to sometimes render all black despite being perfectly fine internally.

TextureCanUseAtlas is best used with lots of small static images. Atlas texture sizes are padded to powers of two, so adding a large image to the atlas makes the atlas use up far more space than necessary. We also use mipmaps with the screenshot textures, so that uses up even more GPU memory with the already massive atlas texture.

I also removed the explicit TextureHasAlphaChannel option because it was redundant.

BUG: 481665