ImagePlatformKWin: Go back to manually combining images
The point of this is to improve image quality with fractional scaling and mixed DPI.
The reason why we aren't just patching the KWin screenshot plugin is because CaptureWorkspace and CaptureArea aren't actually intended for high quality captures.

We are now using OpenCV to scale the images. We do this because OpenCV has higher quality filters and because OpenCV should be fast enough.

This patch should make it so integer scale screenshots are crisper.
Fractional screenshots are already a bit blurry, so this combined with a high quality image filter should be OK.

There's not much more that can be done to fix bug 478426 because of competing requirements.
A combined screens image should have the screens layout, but images need to be scaled to fit the layout and images also need to look as crisp as possible.

BUG: 478426