ImagePlatformKWin: Go back to manually combining images
The point of this is to eventually improve image quality with fractional scaling and mixed DPI. This patch doesn't actually do that yet except that images with an integer scale factor should look nicer with mixed DPI. Maybe we'll use OpenCV or libvips to do higher quality scaling. The reason why we aren't just patching the KWin screenshot plugin is because CaptureWorkspace and CaptureArea aren't actually intended for high quality captures.

Consider this patch a part of a larger change and don't cherry pick this.