SpectacleCore: make class inheritance check more robust
Commit

    f410b0814 "Rename Platform* to ImagePlatform*"

introduced a regression by making spectacle window visible at the
moment of making rectangular screenshot. That happened because the
check for inheritance was using a string instead of a dynamic/qobject
cast, so after the rename the fact the code broke went unnoticed.

Fix that by using qobject_cast() instead, so if rename ever happens
again the code would refuse to compile.

While at it, also add a point that the delay there isn't only because
of the crash but also because X11 compositors different from kwin may
require some time for effects to disappear before the shot can be
made.

BUG: 484652