AnnotationViewport: Copy image section when creating texture instead of setting texture sourceRect
In the rectangle capture mode, annotations were a bit laggy.
Profiling showed that the biggest user of CPU was copying image/texture data while resizing a rectangle annotation.
I assumed that setting the texture sourceRect might be part of why there is so much copying going on.
It turns out I was right, so we copy the section we want before creating the texture instead.
The section changes when the viewportRect changes, but the viewportRect doesn't actually change that often.
This should also help reduce GPU memory usage to a certain degree.
Before, we were making 1 copy of the entire base image and annotation image per screen when there were multiple screens during rectangle capture.