Disable video mode when a new rectangle capture screenshot is started
This was a missing line of code. Without this, doing a rectangle screenshot capture after a video capture causes the fullscreen windows to visually freeze while also being impossible to interact with.

It would be nice if someday we replaced setVideoMode with something that requires less manual management. Manual state management is error prone because it's easy to forget to do it properly.

BUG: 481471