Handle: Fix Qt 6.7 bugs
For some reason, upgrading to Qt 6.7 caused major graphical glitches and binding loops with Handle.
Things that helped fixed the bugs were moving the values out of the ShapePath and into root item properties.
Relying on the root item's width and height directly was particularly buggy for some reason.
I had to add visualWidth and visualHeight properties instead of directly setting implicitWidth and implicitHeight in order to prevent binding loops.