Use custom system tray animated icon
Trying to animate theme icons doesn't work:

- The theme icons use the application's color scheme rather than Plasma's color scheme.
- Real icon theme animation icons such as `process-working` don't actually work. QIcon seems to only support static pixmaps.

Using an animated WebP instead of an animated SVG because you need to use QSvgRenderer to render animated SVGs properly. QMovie does not work with SVG.

Using an animated WebP instead of a GIF because GIF is an old format with lots of drawbacks, like bad alpha support. WebP is just better.

The original Krita project files are included in this commit so that others can edit the icons in the future without starting from scratch.