Improve DialogPage header text
There's no clear "call to action" button on this page, and none of the
screenshot buttons begin with an action verb. So it may not be 100%
obvious how you actually take a screenshot.

This probably won't totally resolve the issue, but we can begin by
changing the headers to make the workflow clearer.