Use Kirigami.NavigationTabBar for immutable toolview tabs
That's what the component was made for, and it's the semantically
correct thing to use here. It looks good and has good usability. It's
more premium than the traditional-style tab bar used in the main
window, and more standard than the custom pseudo segmented button tab
control in the "no screenshot taken automatically" window.