AnnotationDocument: Fix highlighter tool not blending with base image
Previously, the graphics were split into a base image and an annotations image. The change helped to improve CPU performance by not repainting the base image so much, but it messed up the highlighter effect. The highlighter needs to be blended with the base image in order to work correctly.

Now the base image is blended into the annotations image in places where the highlighter is used. This will probably hurt performance somewhat, but it doesn't seem too bad so far.

BUG: 488025

(cherry picked from commit dd2231b5b33cc8d124b68eeeacee19f20cc009b1)

Co-authored-by: Noah Davis <noahadvs@gmail.com>